Cleanup extensions
