+++ Sources/PulseCore/Helpers/CoreData+Extensions.swift
@@ -28,3 +28,40 @@ extension NSManagedObjectContext {
         try count(for: NSFetchRequest<T>(entityName: String(describing: entity)))
     }
 }
+
+extension NSPersistentStoreCoordinator {
+    func createCopyOfStore(at url: URL) throws {
+        guard let sourceStore = persistentStores.first else {
+            throw LoggerStore.Error.unknownError // Should never happen
+        }
+
+        let backupCoordinator = NSPersistentStoreCoordinator(managedObjectModel: managedObjectModel)
+
+        var intermediateStoreOptions = sourceStore.options ?? [:]
+        intermediateStoreOptions[NSReadOnlyPersistentStoreOption] = true
+
+        let intermediateStore = try backupCoordinator.addPersistentStore(
+            ofType: sourceStore.type,
+            configurationName: sourceStore.configurationName,
+            at: sourceStore.url,
+            options: intermediateStoreOptions
+        )
+
+        let backupStoreOptions: [AnyHashable: Any] = [
+            NSReadOnlyPersistentStoreOption: true,
+            // Disable write-ahead logging. Benefit: the entire store will be
+            // contained in a single file. No need to handle -wal/-shm files.
+            // https://developer.apple.com/library/content/qa/qa1809/_index.html
+            NSSQLitePragmasOption: ["journal_mode": "DELETE"],
+            // Minimize file size
+            NSSQLiteManualVacuumOption: true
+        ]
+
+        try backupCoordinator.migratePersistentStore(
+            intermediateStore,
+            to: url,
+            options: backupStoreOptions,
+            withType: NSSQLiteStoreType
+        )
+    }
+}
